refactor: more removal of unused code
