+++ web-client/src/components/MenuDrawer/MenuDrawer.tsx
@@ -1,10 +1,8 @@
 import { Drawer } from 'antd';
 import React from 'react';
-import { useTranslation } from 'react-i18next';
 import { User } from 'src/models/users';
 import styled from 'styled-components';
 
-import { InformationModal, makeLocalStorageKey } from '../Modals/OneTimeModal';
 import SideBottomMenu from '../SideBottomMenu/SideBottomMenu';
 // import SideDrawerProfile from '../SideDrawerProfile/SideDrawerProfile';
 import SideTopMenu from '../SideTopMenu/SideTopMenu';
@@ -15,22 +13,6 @@ const MenuDrawer: React.FC<IMenuDrawerProps> = ({
   profileData,
   logoutHandler,
 }) => {
-  const { t } = useTranslation();
-
-  const instructions = [
-    t('information_modal.MenuDrawer.0'),
-    t('information_modal.MenuDrawer.1'),
-    t('information_modal.MenuDrawer.2'),
-    t('information_modal.MenuDrawer.3'),
-    t('information_modal.MenuDrawer.4'),
-    t('information_modal.MenuDrawer.5'),
-    t('information_modal.MenuDrawer.6'),
-  ];
-  const instructionModalLocalStorageKey = makeLocalStorageKey({
-    prefix: 'reach4help.modalSeen.MenuDrawer',
-    /* TODO:  get unique user id */
-    userid: null,
-  });
   const isLoggedIn = !!(profileData && profileData.displayName);
 
   // Closes SideDrawer if user switches from mobile view
@@ -53,11 +35,6 @@ const MenuDrawer: React.FC<IMenuDrawerProps> = ({
           closeDrawer={closeDrawer}
         />
       </SideDrawer>
-      <InformationModal
-        title={t('information_modal.MenuDrawer.title')}
-        localStorageKey={instructionModalLocalStorageKey}
-        instructions={instructions}
-      />
     </>
   );
 };
