+++ web-client/src/components/SideTopMenu/SideTopMenu.tsx
@@ -13,7 +13,7 @@ import {
 import {
   MyOfferPostsLocationUrl,
   MyRequestPostsLocationUrl,
-} from 'src/modules/requestAndOfferPosts/constants';
+} from 'src/modules/post/constants';
 import { COLORS } from 'src/theme/colors';
 import styled from 'styled-components';
