+++ web-client/src/components/WebClientMap/WebClientMapMarker.tsx
@@ -3,14 +3,14 @@ import React from 'react';
 import { useTranslation } from 'react-i18next';
 import LargeOrangeMarkerIcon from 'src/components/WebClientMap/assets/map-marker-orange-lg.png';
 import SmallOrangeMarkerIcon from 'src/components/WebClientMap/assets/map-marker-orange-sm.png';
-import LargePurpleMarkerIcon from 'src/components/WebClientMap/assets/map-marker-purple-lg.png';
+// TODO: (es) Use thissomewhere? import LargePurpleMarkerIcon from 'src/components/WebClientMap/assets/map-marker-purple-lg.png';
 
-export const OriginMarker: React.FC<OriginMarkerProps> = ({ isCav }) => {
+export const OriginMarker: React.FC<OriginMarkerProps> = () => {
   const { t } = useTranslation();
   return (
     <div style={{ transform: 'translate(-50%, -100%)' }}>
       <img
-        src={isCav ? LargePurpleMarkerIcon : LargeOrangeMarkerIcon}
+        src={LargeOrangeMarkerIcon}
         alt={t('components.web_client_map.a11y_my_location')}
       />
     </div>
@@ -50,8 +50,9 @@ export const SmallOrangeMarker: React.FC<MarkerProps> = () => {
 
 type MarkerProps = Coords;
 
+// TODO: remove?  was previously used for isCav
 interface OriginMarkerProps extends Coords {
-  isCav: boolean;
+  dummy?: boolean;
 }
 
 interface DestinationMarkerProps extends Coords {
