+++ web-client/src/modules/create/containers/CreatePostContainer.tsx
@@ -15,7 +15,7 @@ import NewAddressModal from 'src/modules/create/components/NewAddressModal';
 import PostDetailsStep from 'src/modules/create/components/PostDetailsStep';
 import PostLocationStep from 'src/modules/create/components/PostLocationStep';
 import PostSummary from 'src/modules/create/components/PostSummaryStep';
-import { MyRequestPostsLocationUrl } from 'src/modules/requestAndOfferPosts/constants';
+import { MyRequestPostsLocationUrl } from 'src/modules/post/constants';
 import AuthenticationModal from 'src/pages/modals/AuthenticationModal';
 import { AppState } from 'src/store';
 import styled from 'styled-components';
