+++ web-client/src/modules/personalData/constants.tsx
@@ -1,5 +1,3 @@
 import Location from 'react-app-location';
 
 export const PersonalDataLocation = new Location('/personal-data');
-
-export const RoleInfoLocation = new Location('/personal-data/role-info');
