+++ web-client/src/modules/post/assets/role_pin.png
