+++ web-client/src/modules/post/constants.tsx
