+++ web-client/src/modules/post/containers/MyRequestsContainer.tsx
@@ -1,23 +1,16 @@
 import React, { useEffect } from 'react';
-import { useTranslation } from 'react-i18next';
 import { useDispatch, useSelector } from 'react-redux';
 import { observeMyRequests } from 'src/ducks/MyRequests/actions';
 import { MyRequestsState } from 'src/ducks/MyRequests/types';
 import { ProfileState } from 'src/ducks/profile/types';
-import { ApplicationPreference } from 'src/models/users';
 
 import LoadingWrapper from '../../../components/LoadingComponent/LoadingComponent';
-import {
-  InformationModal,
-  makeLocalStorageKey,
-} from '../../../components/Modals/OneTimeModal';
+import PostList from '../components/GeneralPostList';
 import Header from '../components/Header';
-import PostList from '../components/PublicPostList';
 
 const RequestPostsContainer: React.FC<{ status: string | null }> = ({
   status,
 }) => {
-  const { t } = useTranslation();
   const dispatch = useDispatch();
   const profileState = useSelector(
     ({ profile }: { profile: ProfileState }) => profile,
@@ -41,36 +34,12 @@ const RequestPostsContainer: React.FC<{ status: string | null }> = ({
     return <LoadingWrapper />;
   }
 
-  const instructions = [
-    t('information_modal.AcceptedRequestsContainer.0'),
-    t('information_modal.AcceptedRequestsContainer.1'),
-  ];
-  const instructionModalLocalStorageKey = makeLocalStorageKey({
-    prefix: 'reach4help.modalSeen.AcceptedRequestsContainer',
-    userid: profileState.uid,
-  });
-
   return (
     <>
-      <Header
-        requestsType={t(
-          'modules.requests.containers.AcceptedRequestContainer.accepted',
-        )}
-        numRequests={Object.keys(myRequests.data || {}).length}
-        isCav={
-          profileState.profile?.applicationPreference ===
-          ApplicationPreference.cav
-        }
-        isAcceptedRequests
-      />
+      <Header numRequests={Object.keys(myRequests.data || {}).length} />
       <PostList
         posts={myRequests.data}
         loading={myRequests && myRequests.loading}
-      />
-      <InformationModal
-        title={t('information_modal.AcceptedRequestsContainer.title')}
-        localStorageKey={instructionModalLocalStorageKey}
-        instructions={instructions}
       />
     </>
   );
