+++ web-client/src/pages/MasterPage.tsx
@@ -10,7 +10,7 @@ import {
 } from 'react-router-dom';
 import DashboardLayout from 'src/components/DashboardLayout/DashboardLayout';
 import { signOutCurrentUserAction } from 'src/ducks/auth/actions';
-import { RootUrl } from 'src/modules/requestAndOfferPosts/constants';
+import { RootUrl } from 'src/modules/post/constants';
 import { Module } from 'src/types/module';
 
 import modules from '../modules';
@@ -37,13 +37,6 @@ const MasterPage = (): ReactElement => {
       case '/personal-data/role-info':
       case '/phone/entry':
       case '/phone/verify':
-      case `${RootUrl}/accepted`:
-      case `${RootUrl}/archived`:
-      case `${RootUrl}/find`:
-      case `${RootUrl}/finished`:
-      case `${RootUrl}/new`:
-      case `${RootUrl}/ongoing`:
-      case `${RootUrl}/open`:
         return `${title}: `.concat(
           t(`routeSubtitles.${path.replace(/\//g, '_')}`),
         );
