+++ web-client/src/pages/routes/ProtectedRoute.tsx
@@ -4,10 +4,7 @@ import { Redirect, Route, RouteProps, useLocation } from 'react-router-dom';
 import { observeUserAction } from 'src/ducks/auth/actions';
 import { observePrivileged, observeProfile } from 'src/ducks/profile/actions';
 import { ProfileState } from 'src/ducks/profile/types';
-import {
-  PersonalDataLocation,
-  RoleInfoLocation,
-} from 'src/modules/personalData/constants';
+import { PersonalDataLocation } from 'src/modules/personalData/constants';
 import { AppState } from 'src/store';
 
 import LoadingWrapper from '../../components/LoadingComponent/LoadingComponent';
@@ -69,17 +66,6 @@ const ProtectedRoute: React.FC<RouteProps> = ({ path, component }) => {
       <Redirect
         to={{
           pathname: PersonalDataLocation.path,
-          state: { redirectBack: `${location.pathname}${location.search}` },
-        }}
-      />
-    );
-  }
-
-  if (!profileState.profile.applicationPreference) {
-    return (
-      <Redirect
-        to={{
-          pathname: RoleInfoLocation.path,
           state: { redirectBack: `${location.pathname}${location.search}` },
         }}
       />
